refactor(node): refactor `OutputOptions` using zod
