+++ packages/rolldown/src/options/normalized-input-options.ts
@@ -1,7 +1,4 @@
-import type {
-  LogHandler,
-  NormalizedInputOptions as RollupNormalizedInputOptions,
-} from '../rollup'
+import type { NormalizedInputOptions as RollupNormalizedInputOptions } from '../rollup'
 import type { InputOptions } from './input-options'
 import { Plugin, ParallelPlugin } from '../plugin'
